+++ glue_vispy_viewers/volume/tests/test_volume_viewer.py
@@ -7,10 +7,19 @@ from glue.core import DataCollection, Data
 from glue.app.qt.application import GlueApplication
 from glue.core.component import Component
 from glue.core.link_helpers import LinkSame
+from glue.core.fixed_resolution_buffer import PIXEL_CACHE, ARRAY_CACHE
 
 from ..volume_viewer import VispyVolumeViewer
 
 GLUE_LT_08 = LooseVersion(glue.__version__) < LooseVersion('0.8')
+
+
+def teardown_function(function):
+    # Make sure cache is empty
+    if len(PIXEL_CACHE) > 0:
+        raise Exception("Pixel cache contains {0} elements".format(len(PIXEL_CACHE)))
+    if len(ARRAY_CACHE) > 0:
+        raise Exception("Array cache contains {0} elements".format(len(ARRAY_CACHE)))
 
 
 def make_test_data(dimensions=(10, 10, 10)):